refactor: isolate UI trace artifact saving
